Apply default network bandwidth to tasks if not provided.
We make network bandwidth a first-class resource in Mesos. Therefore,
every task should have it declared when falling on a slave declaring
network bandwidth.
If task do not declare the reserved amount, Mesos reserves a default
amount taken out from the unreserved resources and related to
the share of reserved CPU.

In order to support every frameworks not supporting network bandwidth
yet, the resource can also be declared with a label called
NETWORK_BANDWIDTH_RESOURCE.

Note: This implementation only distributes unreserved resources. This
means that in nominal case, the scheduler can match offer for any role
if support for network bandwidth is correctly implemented. However, in
the case of labels and default value, the network bandwidth is
automatically taken from unreserved pool.