Prevent Mesos from killing container based on oom notification.
Linux Kernel oom notification doesn't necessarily mean that oom-killer
has been invoked, and when OOM is reached because of page cache usage,
we'd probably like to keep the container alive and give it a chance to
reclaim page cache memory if needed. If rss reaches the limit, the
Kernel will still kill the most memory consuming process, which should
probably lead to Mesos killing the container.